Konverse_API/requirements.txt
flask
google.generativeai
peewee
tidb-vector
pypdf
flask-cors